feat: refactor proposal table
