+++ src/lib/components/table/proposals/ProposalTextCell.tsx
