+++ src/lib/components/table/proposals/ProposalsTableHeader.tsx
@@ -1,20 +1,31 @@
-import type { GridProps } from "@chakra-ui/react";
-import { Grid } from "@chakra-ui/react";
+import type { DividerProps, GridProps } from "@chakra-ui/react";
+import { Grid, Text } from "@chakra-ui/react";
 
-import { TableHeader } from "../tableComponents";
+import { TableHeader, TableHeaderFreeze } from "../tableComponents";
 
 export const ProposalsTableHeader = ({
   templateColumns,
+  boxShadow,
 }: {
   templateColumns: GridProps["templateColumns"];
-}) => (
-  <Grid templateColumns={templateColumns}>
-    <TableHeader>Proposal ID</TableHeader>
-    <TableHeader>Proposal Title</TableHeader>
-    <TableHeader textAlign="center">Status</TableHeader>
-    <TableHeader>Voting ends</TableHeader>
-    <TableHeader>Resolved Block Height</TableHeader>
-    <TableHeader>Type</TableHeader>
-    <TableHeader>Proposer</TableHeader>
-  </Grid>
-);
+  boxShadow: DividerProps["boxShadow"];
+}) => {
+  // TODO - Revisit split columnsWidth
+  const columnsWidth = templateColumns?.toString().split(" ");
+  return (
+    <Grid templateColumns={templateColumns} minW="min-content">
+      <TableHeaderFreeze left="0">Proposal ID</TableHeaderFreeze>
+      <TableHeaderFreeze
+        left={columnsWidth && columnsWidth[0]}
+        boxShadow={boxShadow}
+        color="gray.800"
+      >
+        <Text color="text.main">Proposal Title/Type</Text>
+      </TableHeaderFreeze>
+      <TableHeader textAlign="center">Status</TableHeader>
+      <TableHeader>Voting ends</TableHeader>
+      <TableHeader>Resolved Block Height </TableHeader>
+      <TableHeader>Proposed By </TableHeader>
+    </Grid>
+  );
+};
