+++ src/lib/components/table/proposals/ProposalsTableRow.tsx
@@ -1,12 +1,17 @@
-import type { GridProps } from "@chakra-ui/react";
-import { Grid, Text } from "@chakra-ui/react";
+import type { DividerProps, GridProps } from "@chakra-ui/react";
+import { Grid } from "@chakra-ui/react";
 
-import { TableRow } from "../tableComponents";
+import { TableRow, TableRowFreeze } from "../tableComponents";
+import { useCelatoneApp } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
+import { StopPropagationBox } from "lib/components/StopPropagationBox";
 import { Proposer } from "lib/components/table/proposals/Proposer";
-import type { Proposal } from "lib/types";
+import { openNewTab } from "lib/hooks";
+import { AmpTrackMintscan } from "lib/services/amplitude";
+import type { Option, Proposal } from "lib/types";
 import { ProposalStatus } from "lib/types";
 
+import { ProposalTextCell } from "./ProposalTextCell";
 import { ResolvedHeight } from "./ResolvedHeight";
 import { StatusChip } from "./StatusChip";
 import { VotingEndTime } from "./VotingEndTime";
@@ -14,29 +19,77 @@ import { VotingEndTime } from "./VotingEndTime";
 export interface ProposalsTableRowProps {
   proposal: Proposal;
   templateColumns: GridProps["templateColumns"];
+  boxShadow: DividerProps["boxShadow"];
 }
 
 export const ProposalsTableRow = ({
   proposal,
   templateColumns,
+  boxShadow,
 }: ProposalsTableRowProps) => {
+  const {
+    chainConfig: {
+      explorerLink: { proposal: explorerProposal },
+    },
+  } = useCelatoneApp();
+
+  // TODO - Revisit split columnsWidth
+  const columnsWidth = templateColumns?.toString().split(" ");
   const isDepositFailed = proposal.status === ProposalStatus.DEPOSIT_FAILED;
   const isDepositOrVoting =
     proposal.status === ProposalStatus.DEPOSIT_PERIOD ||
     proposal.status === ProposalStatus.VOTING_PERIOD;
+
+  const hoverBg = (): Option<string> => {
+    if (proposal.isExpedited && isDepositOrVoting) return "primary.background";
+    return isDepositFailed ? undefined : "gray.900";
+  };
+
   return (
-    <Grid templateColumns={templateColumns}>
-      <TableRow>
+    <Grid
+      templateColumns={templateColumns}
+      minW="min-content"
+      cursor={isDepositFailed ? "default" : "pointer"}
+      _hover={{ "> div": { bgColor: hoverBg } }}
+      onClick={
+        !isDepositFailed
+          ? () => {
+              AmpTrackMintscan("proposal-detail", {
+                type: proposal.type,
+                status: proposal.status,
+              });
+              openNewTab(
+                `${explorerProposal}/${proposal.proposalId.toString()}`
+              );
+            }
+          : undefined
+      }
+    >
+      <TableRowFreeze left="0">
         <ExplorerLink
           isReadOnly={isDepositFailed}
           type="proposal_id"
           value={proposal.proposalId.toString()}
           showCopyOnHover
+          ampCopierSection="proposal-list"
         />
-      </TableRow>
-      <TableRow>{proposal.title}</TableRow>
+      </TableRowFreeze>
+      <TableRowFreeze
+        left={columnsWidth && columnsWidth[0]}
+        boxShadow={boxShadow}
+        color="gray.800"
+      >
+        <ProposalTextCell
+          title={proposal.title}
+          type={proposal.type}
+          isExpedited={proposal.isExpedited}
+          isDepositOrVoting={isDepositOrVoting}
+        />
+      </TableRowFreeze>
       <TableRow justifyContent="center">
-        <StatusChip status={proposal.status} />
+        <StopPropagationBox>
+          <StatusChip status={proposal.status} />
+        </StopPropagationBox>
       </TableRow>
       <TableRow>
         <VotingEndTime
@@ -46,16 +99,19 @@ export const ProposalsTableRow = ({
         />
       </TableRow>
       <TableRow>
-        <ResolvedHeight
-          resolvedHeight={proposal.resolvedHeight}
-          isDepositOrVoting={isDepositOrVoting}
+        <StopPropagationBox>
+          <ResolvedHeight
+            resolvedHeight={proposal.resolvedHeight}
+            isDepositOrVoting={isDepositOrVoting}
+            amptrackSection="proposal-list"
+          />
+        </StopPropagationBox>
+      </TableRow>
+      <TableRow>
+        <Proposer
+          proposer={proposal.proposer}
+          amptrackSection="proposal-list"
         />
-      </TableRow>
-      <TableRow>
-        <Text color="text.dark">{proposal.type}</Text>
-      </TableRow>
-      <TableRow>
-        <Proposer proposer={proposal.proposer} />
       </TableRow>
     </Grid>
   );
